strategies/edgelab/requirements.txt
aiohttp==3.9.5
websockets==12.0
redis==5.0.8
aiosqlite==0.20.0
streamlit==1.38.0
pandas==2.2.2
numpy==1.26.4
plotly==5.22.0
rapidfuzz==3.9.4
python-dotenv==1.0.1
requests==2.32.3
httpx==0.27.0
python-telegram-bot==21.5
pytest==8.2.2
pytest-asyncio==0.23.7
web3>=6.0.0
APScheduler==3.10.4
